streamline plugin installation process
