+++ drizzle_migrations/20250829134938_bright_thunderbolts.sql
@@ -330,7 +330,7 @@ CREATE TABLE "plugins" (
 	"id" uuid PRIMARY KEY NOT NULL,
 	"plugin_id" text NOT NULL,
 	"is_activated" boolean DEFAULT false NOT NULL,
-	"is_installed" boolean DEFAULT true NOT NULL,
+	"is_installed" boolean DEFAULT false NOT NULL,
 	"backup" boolean DEFAULT false NOT NULL,
 	"created_at" timestamp (3) with time zone DEFAULT now() NOT NULL,
 	"updated_at" timestamp (3) with time zone DEFAULT now(),
@@ -489,6 +489,7 @@ CREATE TABLE "venue_bookings" (
 );
 --> statement-breakpoint
 CREATE TABLE "venues" (
+	"capacity" integer,
 	"created_at" timestamp (3) with time zone DEFAULT now() NOT NULL,
 	"creator_id" uuid,
 	"description" text,
