+++ drizzle_migrations/meta/20250829134938_snapshot.json
@@ -1,5 +1,5 @@
 {
-  "id": "12fc2106-e390-4ba4-bdbc-cc05423baa43",
+  "id": "4359d958-9015-4de2-997d-53ebc22e3385",
   "prevId": "00000000-0000-0000-0000-000000000000",
   "version": "7",
   "dialect": "postgresql",
@@ -4859,7 +4859,7 @@
           "type": "boolean",
           "primaryKey": false,
           "notNull": true,
-          "default": true
+          "default": false
         },
         "backup": {
           "name": "backup",
@@ -7164,6 +7164,12 @@
       "name": "venues",
       "schema": "",
       "columns": {
+        "capacity": {
+          "name": "capacity",
+          "type": "integer",
+          "primaryKey": false,
+          "notNull": false
+        },
         "created_at": {
           "name": "created_at",
           "type": "timestamp (3) with time zone",
